Finalising the first draft of the blockBreakingSystem
Adds a command to register Blocks, being part of the woolbattle map, to differentiate between wool being placed in the game and wool, being a fragment  of these very logistics. The wool, the players breaking it are granted now is coloured according to the players team, which is defined manually in source by a variable.