Pushing the miraculously working version of the plugin
The idea folders explicitly are intended to be able to be found in this place, as they perhaps are conditioning the previously documented effect. If the occasion of a testimony of the functionality of the plugin in another hardware-dependent location will occur, if my fellow coprogrammers, or me myself will be able to document, or classify the effect, these fractions of the current version are going to be removed. I personally  consider myself an amateur in the sphere of programming, maven displays, which could be about to condition the correction of an error, obvious to the practiced eye.