Adding an implementation of  the mongodb java api
Through this very patch the functionality of storing blocks of a given woolbattle map in mongodb is possessing actual functionality for the first time. The efforts undertaken by me, to trace down the initial error (confusing the indexes of an array of Doubles, in order to fetch them to a local array of the blocks belonging to the map that made the implemenation seem infunctional) have summited in this very  release and I, whilst considering my poor understanding of the sphere of programming in generaly, somewhat proudly present the first draft of the system, I tried to create through my tendencies. Additionally to that a conversion method of  an instance of an ArrayList, with an org.Bukkit.Location generic, towards a one, possessing an ArrayList<Double> generic is implemented, is about to be used in order to abbreviate some code cascades, undertaken by me preciously. The over all asthetics may not be sufficient after all to be implemented in the contingent of functionalities, provided by my fellow developers on other branches, but a major improvement has been achieved, by means of the functionality of the plugin's fragment, being developed by me and my very sanity. Polishing will be a consequence to this version of the Woolsystem, considering the poor implementation of these functionalities. Another commit without doubts will condition said processes.