Passive perks patch, fix of a non-working double jump after an act of changing game modes
Fixes the previously infunctional system of passive perk through a correction of a grievance, introduced in the act of administering data concerning players fetched from the db.
Corrects additionally to that the issue of the movement capability of the double jump not working after a modification made to a players game mode.
Further cleans minor redundancies in the code.